salesdata/requirement.txt
crewai==0.98.0
streamlit==1.42.0
matplotlib==3.10.0
seaborn==0.13.2
psycopg2-binary==2.9.10
pyodbc==5.2.0
google_auth_oauthlib==1.2.1
google-api-python-client==2.160.0
google-generativeai==0.8.4
pymssql==2.3.2